requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
Jinja2==3.0.1
PyPDF2==3.0.0
nltk==3.6.3
python-dotenv==0.19.1
gunicorn==20.1.0 